refactor: GunComponent 碎片化，並移除舊的槍支組件
