+++ scripts/modules/gun_system/components/Component.ts
@@ -1,13 +1,19 @@
+import { GunComponent } from "./gun/GunComponent";
+import { GunRecoilComponent } from "./gun/GunAnimationComponent";
+import { GunFireComponent } from "./gun/GunFireComponent";
+import { GunReloadComponent } from "./gun/GunReloadComponent";
 import { EntityComponent } from "./EntityComponent";
 import { BulletComponent } from "./BulletComponent";
 import { DamageComponent } from "./DamageComponent";
-import { GunComponent } from "./GunComponent";
 import { ItemComponent } from "./ItemComponent";
 import { MagazineComponent } from "./MagazineComponent";
 
 export type ComponentTypes = {
-    'damage': DamageComponent;
     'gun': GunComponent;
+    'gun_fire': GunFireComponent
+    'gun_recoil': GunRecoilComponent;
+    'gun_reload': GunReloadComponent;
+    'damage': DamageComponent;
     'item': ItemComponent;
     'magazine': MagazineComponent;
     'entity': EntityComponent;
