+++ scripts/modules/gun_system/components/gun/GunComponent.ts
@@ -1,0 +1,18 @@
+import { FireModeEnum, GunTypeEnum } from "../../types/Enums";
+import { Component } from "../Component";
+
+type ComponentDataType = {
+    gunTypeEnum: GunTypeEnum;
+}
+
+export class GunComponent extends Component {
+    
+    readonly gunTypeId: GunTypeEnum;
+
+    constructor(data: ComponentDataType) {
+        super('gun');
+
+        this.gunTypeId = data.gunTypeEnum;
+    }
+
+}
